This is revision 08.12, (pre) Kerniger Kaernter
Outline:
* First, incomplete info.xml version.
* The .deb file is eliminated. Most items are not converted to LOG_DBG() yet
* Up to date SIMP optimization ((multi) frequent piezo with acoustics)
* Ersatz material optimization from Bastian Schmidt
* Compressed textual output 
* VTK export but options to be set in ...SimInOut/VTK/CMakeLists.txt!

Details:
* Removed XML/TestParser
* New schema structure for optimization
* The default saveBegin is now step 0 as the 0th optimization iteration
  is the start configuration
* Schame change: the output formats need no id, it is by default the name
* Added VTK output. It is an option for the ParaView reader plugin
* Transient analysis: you can give a frequency list 
* Now a (incomplete) info.xml file is written. It can be postprocessed
  via XSL (e.g. auxiliary/XSL/extract_eigenvalues_line.xsl):
  xsltproc <stylesheed>.xsl <problem>.info.xml > gnuplot.dat
* The technique is InfoNode, specialization to ParamNode with
  a global "info" pointer. On several occations (precond, solver, ...)
  the current version of the info.xml is written.
* new and changed program options (version, history)
* Ensight is removed
* Assemble::AssembleLinRHS() has no actTime parameter any more as it
  was not used.
* There is no transientharmonic any more
* OLAS has now a InitRHS(Vector<Double>*)

+++ source/OLAS/matvec/sbmmatrix.cc
@@ -99,7 +99,7 @@ namespace OLAS {
     if ( myBlockSize_ == 0 ) {
       myBlockSize_ = blockSize;
     }
-    else if ( blockSize != myBlockSize_ ) {
+    else if ( (UInt) blockSize != myBlockSize_ ) {
       (*error) << "SBM_Matrix::SetSubMatrix: Block size " << blockSize
                << " does not match block size " << myBlockSize_
                << " of other sub-matrices";
@@ -108,7 +108,7 @@ namespace OLAS {
 
     // Check that the dimensions of the new matrix fit to its neighbours
     // in the super-block matrix
-    UInt auxDim;
+    Integer auxDim;
 
     // left neighbour
     if ( j > 1 && subMat_[ComputeIndex(i,j-1)] != NULL ) {
@@ -191,7 +191,7 @@ namespace OLAS {
       StdMatrix *auxMat = NULL;
 
       // Loop over all sub-blocks and trigger addition
-      for ( UInt k = 1; k <= nrows_ * ncols_; k++ ) {
+      for ( Integer k = 1; k <= nrows_ * ncols_; k++ ) {
 
         // Get hold of submatrix
         auxMat = sbmMat.subMat_[k];
@@ -558,7 +558,7 @@ namespace OLAS {
     Integer sum = 0;
     StdMatrix *stdMat = NULL;
 
-    for ( UInt i = 1; i <= nrows_; i++ ) {
+    for ( Integer i = 1; i <= nrows_; i++ ) {
       stdMat =  subMat_[ ComputeIndex(i,i) ];
       if ( stdMat != NULL ) {
         sum += stdMat->GetNrowsScalar();
@@ -584,7 +584,7 @@ namespace OLAS {
     Integer sum = 0;
     StdMatrix *stdMat = NULL;
 
-    for ( UInt i = 1; i <= ncols_; i++ ) {
+    for ( Integer i = 1; i <= ncols_; i++ ) {
       stdMat =  subMat_[ ComputeIndex(i,i) ];
       if ( stdMat != NULL ) {
         sum += stdMat->GetNcolsScalar();
